Removing cast

+++ src/stan/services/variational/print_progress.hpp
@@ -38,8 +38,8 @@ namespace stan {
                                    "Total number of iterations",
                                    m);
         stan::math::check_nonnegative(function,
-                                   "Starting iteration",
-                                   start);
+                                      "Starting iteration",
+                                      start);
         stan::math::check_positive(function,
                                    "Final iteration",
                                    finish);
@@ -54,7 +54,7 @@ namespace stan {
           o << std::setw(it_print_width) << m + start
             << " / " << finish;
           o << " [" << std::setw(3)
-            << static_cast<int>( (100.0 * (start + m)) / finish )
+            << (100 * (start + m)) / finish
             << "%] ";
           o << (tune ? " (Adaptation)" : " (Variational Inference)");
           o << suffix;
